refactored metrics entrypoint
